+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # C6UI
 
 ## v3.9.0 (June 9, 2016)
-* [FEATURE]: Add _error flag to c6Db model if a save, refresh or erase
+* [FEATURE]: Add \_error flag to c6Db model if a save, refresh or erase
   request fails
 
 ## v3.8.0 (June 8, 2016)